+++ src/FinServ.Application/Handlers/Clientes/GetAtivos/GetAtivosResponse.cs
@@ -5,9 +5,9 @@
         public string NomeProduto { get; set; }
         public string DescricaoProduto { get; set; }
         public string DataVencimentoProduto { get; set; }
-        public double TaxaJurosMes { get; set; }
-        public double ValorCompra { get; set; }
-        public double ValorAtual { get; set; }
+        public decimal TaxaJurosMes { get; set; }
+        public decimal ValorCompra { get; set; }
+        public decimal ValorAtual { get; set; }
         public int Quantidade { get; set; }
         public string DataCompra { get; set; }
