+++ src/FinServ.Application/Handlers/Produtos/QueryProdutoByCodigo/QueryProdutoByCodigoResponse.cs
@@ -6,7 +6,7 @@ namespace FinServ.Application.Handlers.Produtos.QueryProdutoByCodigo
     {
         public int IdProduto { get; set; }
         public string Nome { get; set; }
-        public double TaxaJurosMensal { get; set; }
+        public decimal TaxaJurosMensal { get; set; }
         public DateTime DataVencimento { get; set; }
         public int Quantidade { get; set; }
         public int CodigoProduto { get; set; }
